requirements.txt
pytelegrambotapi==4.4.0
python-dotenv==0.19.2
wikipedia==1.4.0
